LayerDM/tests/packages/build_module/pyproject.toml
[build-system]
requires = [
  "scikit-build-core",
  "vtk-sdk-python-wheel-helper",
  "vtk-sdk==9.6.0",
  "slicer-core-sdk==5.11.0.*",
  "slicer-layer-dm-sdk"
]
build-backend = "scikit_build_core.build"

[project]
name = "slicer_layer_dm_pipeline"
version = "1.0.0"
authors = [
  { name = "Alexy Pellegrini", email = "alexy.pellegrini@kitware.com" }
]
dependencies = ["slicer-layer-dm"]
